requirements.txt
fastapi
uvicorn
requests
langchain
faiss-cpu
pypdf
python-dotenv
langchain-community
langchain-google-genai
